+++ src/BusinessLogic/XMLParser.java
@@ -43,10 +43,10 @@ public class XMLParser {
 					foods.add(f);
 				}
 				NodeList rawMasses = mealelem.getElementsByTagName("Mass");
-				ArrayList<Integer> masses = new ArrayList<Integer>();
+				ArrayList<Double> masses = new ArrayList<Double>();
 				for (int j = 0; j < rawMasses.getLength(); j++) {
 					Element masselem = (Element)rawMasses.item(j);
-					int mass = Integer.parseInt(masselem.getTextContent());
+					double mass = Double.parseDouble(masselem.getTextContent());
 					masses.add(mass);
 				}
 
@@ -131,7 +131,7 @@ public class XMLParser {
 				}
 				Element masses = doc.createElement("Masses");
 				meal.appendChild(masses);
-				for (int i: m.getMasses()) {
+				for (double i: m.getMasses()) {
 					Element mass = doc.createElement("Mass");
 					mass.appendChild(doc.createTextNode("" + i));
 					masses.appendChild(mass);
@@ -186,10 +186,10 @@ public class XMLParser {
 								foods.add(f);
 							}
 							NodeList rawMasses = timeElem.getElementsByTagName("Mass");
-							ArrayList<Integer> masses = new ArrayList<Integer>();
+							ArrayList<Double> masses = new ArrayList<Double>();
 							for (int l = 0; l < rawMasses.getLength(); l++) {
 								Element masselem = (Element)rawMasses.item(l);
-								int mass = Integer.parseInt(masselem.getTextContent());
+								double mass = Double.parseDouble(masselem.getTextContent());
 								masses.add(mass);
 							}
 
@@ -260,7 +260,7 @@ public class XMLParser {
 							}
 							Element masses = doc.createElement("Masses");
 							meal.appendChild(masses);
-							for (int h: m.getMasses()) {
+							for (double h: m.getMasses()) {
 								Element mass = doc.createElement("Mass");
 								mass.appendChild(doc.createTextNode("" + h));
 								masses.appendChild(mass);
